Finesse and Publish/museum-kiosk-pipeline-project/historical_data_pipeline/requirements.txt
boto3
pytest
pylint
python-dotenv
ipykernel